+++ py/sdk/v3/__init__.py
@@ -6,6 +6,7 @@ from .graphs import *
 from .indices import *
 from .prompts import *
 from .retrieval import *
+from .system import *
 from .users import *
 
 __all__ = [
@@ -17,5 +18,6 @@ __all__ = [
     "Indices",
     "Prompts",
     "Retrieval",
+    "System",
     "Users",
 ]
